requirements.txt
Flask==1.0.2
pandas==0.25.3
google-cloud-datastore==1.12.0
google-api-core==1.17.0
